Extract the daemon supervisor into a new daemon workspace crate
crates/daemon-internal (package name: daemon) now owns the daemon process
body: the supervised generation loop (in-process restart on a namespace
change, flap backstop, port-stuck fallback), the zenoh router host and its
watchdog, the core-node runner, the router-federation lifecycle with its
UDS control socket, and the two CLI<->daemon shared surfaces (the on-disk
DaemonState handoff file, moved to state.rs, and the control-socket
protocol plus blocking poke client, moved to control.rs).

The public API is serve(ServeOptions) plus a plain ClockSource; the clap
ValueEnum stays in peppy and converts via From. The binary's git hash is
passed in as data (PEPPY_GIT_HASH stays read by peppy's build.rs), so the
library reads no build-time env. state_file_in loses its test-support gate,
keeping the daemon crate feature-free.

peppy's service serve command becomes a thin wrapper; its error enum drops
MissingMessagingRouter (moved) and the dead UnsupportedEngine,
MissingEngineConfig, and Zenohd variants, and gains a converting Daemon
variant. pmi drops build_zenoh in peppy (kept transitively via daemon), and
the Compact log filter widens to peppy,daemon,auth so daemon-side logs are
not silently dropped in CLI runs. tests/service_serve.rs and the e2e
lifecycle test are unchanged.
