+++ crates/daemon-internal/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "daemon"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+
+[dependencies]
+auth = { workspace = true }
+config = { workspace = true }
+core-node = { workspace = true }
+daemon-config = { workspace = true }
+json5-pretty = { workspace = true }
+pmi = { workspace = true, features = ["zenoh", "build_zenoh"] }
+# Safe wrappers over the libc syscalls the daemon needs (the between-generations
+# waitpid(WNOHANG) reaping and the kill(pid, 0) DaemonState liveness probe), so
+# the crate carries no `unsafe` for them.
+rustix = { version = "1", features = ["process"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+serde_json5 = "0.2"
+thiserror = "2.0"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal", "sync", "net", "io-util", "time"] }
+tokio-util = "0.7"
+tracing = "0.1"
+
+[dev-dependencies]
+tempfile = "3.8"
