+++ crates/daemon-internal/src/shutdown_signal.rs
