+++ crates/peppy/Cargo.toml
@@ -6,7 +6,8 @@ autotests = false
 
 [dependencies]
 auth = { workspace = true }
-pmi = { workspace = true, features = ["zenoh", "build_zenoh"] }
+daemon = { workspace = true }
+pmi = { workspace = true, features = ["zenoh"] }
 config = { workspace = true }
 daemon-config = { workspace = true }
 containers = { workspace = true }
@@ -31,8 +32,8 @@ tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 derive_more = { version = "2", features = ["from"] }
 service-manager = "0.11"
-# Safe wrappers over the libc syscalls peppy needs (getuid/geteuid and the
-# kill(pid, 0) liveness probe), so the crate carries no `unsafe` for them.
+# Safe wrappers over the libc syscalls peppy needs (getuid/geteuid in
+# `service install`), so the crate carries no `unsafe` for them.
 rustix = { version = "1", features = ["process"] }
 dirs = "6"
 names-generator2 = "0.2.1"
