+++ crates/peppy/src/error.rs
@@ -9,14 +9,12 @@ pub enum Error {
     #[from]
     Io(std::io::Error),
 
-    // -- serve
-    UnsupportedEngine,
-    MissingEngineConfig,
-    MissingMessagingRouter,
+    // -- daemon: errors surfaced by the `daemon` crate (`peppy service serve`)
+    #[from]
+    Daemon(daemon::DaemonError),
 
     // -- commands
     ExecutionFailed(String),
-    Zenohd(String),
     Sync(String),
 
     // -- Node
@@ -49,15 +47,12 @@ impl Display for Error {
     fn fmt(&self, fmt: &mut Formatter) -> core::result::Result<(), core::fmt::Error> {
         match self {
             Error::Io(e) => write!(fmt, "IO error: {e}"),
-            Error::UnsupportedEngine => write!(fmt, "Unsupported engine"),
-            Error::MissingEngineConfig => write!(fmt, "Missing engine config"),
-            Error::MissingMessagingRouter => write!(fmt, "Missing messaging router"),
+            Error::Daemon(e) => write!(fmt, "{e}"),
             Error::ExecutionFailed(msg) => {
                 // Convert escaped newlines to actual newlines for readable output
                 let readable_msg = msg.replace("\\n", "\n");
                 write!(fmt, "{readable_msg}")
             }
-            Error::Zenohd(msg) => write!(fmt, "Zenohd error: {msg}"),
             Error::Sync(msg) => write!(fmt, "Sync error: {msg}"),
             Error::InvalidNodeName(name) => write!(fmt, "Invalid node name: {name}"),
             Error::NodeWatcher(msg) => write!(fmt, "Node watcher error: {msg}"),
@@ -96,19 +91,6 @@ mod tests {
         assert_eq!(
             Error::Io(std::io::Error::other("boom")).to_string(),
             "IO error: boom"
-        );
-        assert_eq!(Error::UnsupportedEngine.to_string(), "Unsupported engine");
-        assert_eq!(
-            Error::MissingEngineConfig.to_string(),
-            "Missing engine config"
-        );
-        assert_eq!(
-            Error::MissingMessagingRouter.to_string(),
-            "Missing messaging router"
-        );
-        assert_eq!(
-            Error::Zenohd("x".to_string()).to_string(),
-            "Zenohd error: x"
         );
         assert_eq!(Error::Sync("x".to_string()).to_string(), "Sync error: x");
         assert_eq!(
